Exception handler for HttpNotReadableException exceptions implemented (thrown for example when there's an unknown property in the body).
For an exception to be thrown when there's an unknown property in the body, Jackson deserialization configuration 'fail-on-unknown-properties' had to be set to true on application.yml

AuthenticationInterceptor now sends Problem object in case of missing token or token not being bearer token. The problem type for these is now "authentication".

Implemented tests for the entirety of UsersService.